requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg[binary]
Flask-CORS==4.0.0

# Added security libs
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Added validation lib
marshmallow==3.20.2
